rename folder boardin to onBoarding
