+++ src/screens/onBoarding/style.ts
